[refactor]重構路由以使用控制器
